Compatibility updates
These changes allow Steppy 2 to compile with Xcode 7.2 and Swift 2.1.1
. However, it will not be published as an update on the App Store as it
is still missing some components like extensions and the watch app. It
is also not very “clean” and lacks usage of the latest swift features.
Because of this it will be kept in a separate branch from the master
which will remain to be the App Store version.